fix: move ToStringKeyValues to default_adapter to fix import cycle
